LLM_CSV_DATA_ANALYSIS/requirements.txt
langchain==0.1.13
streamlit==1.29.0
openai==1.14.2
tiktoken==0.5.2
python-dotenv==1.0.1
langchain-openai==0.1.0
langchain_experimental==0.0.55
tabulate==0.9.0